+++ src/pages/index.js
@@ -1,6 +1,7 @@
-export { Analyze } from './Analyze/Analyze';
-export { Contact } from './Contact';
-export { Future } from './Future';
-export { Home } from './Home';
-export { Music } from './Music';
-export { Projects } from './Projects';
+export { Analyze } from "./Analyze/Analyze";
+export { Contact } from "./Contact";
+export { Future } from "./Future";
+export { Home } from "./Home";
+export { Music } from "./Music";
+export { Projects } from "./Projects";
+export { Resume } from "./Resume";